Moves There is a Take created step implementation
